refactor: review type 분리
